OpenDAL

Description

This module provides Haskell bindings for OpenDAL.

Documentation

data Operator

Operator is the entry for all public blocking APIs. Create an Operator with newOp.

data Lister

Lister is designed to list entries at given path in a blocking manner. Users can construct Lister by listOp or scanOp.

data OpenDALError

Represents an error that can occur when using OpenDAL.

Constructors

OpenDALError

Fields

errorCode :: ErrorCode

The error code.

message :: String

The error message.

data ErrorCode

Represents the possible error codes that can be returned by OpenDAL.

Constructors

FFIError
An error occurred in the FFI layer.

Unexpected
OpenDAL don't know what happened here, and no actions other than just returning it back. For example, s3 returns an internal service error.

Unsupported
Underlying service doesn't support this operation.

ConfigInvalid
The config for backend is invalid.

NotFound
The given path is not found.

PermissionDenied
The given path doesn't have enough permission for this operation.

IsADirectory
The given path is a directory.

NotADirectory
The given path is not a directory.

AlreadyExists
The given path already exists thus we failed to the specified operation on it.

RateLimited
Requests that sent to this path is over the limit, please slow down.

IsSameFile
The given file paths are same.

data EntryMode

Represents the mode of an entry in a storage system (e.g., file or directory).

Constructors

File

Dir

Unknown

data Metadata

Represents metadata for an entry in a storage system.

Constructors

Metadata

Fields

mMode :: EntryMode

The mode of the entry.

mCacheControl :: Maybe String

The cache control of the entry.

mContentDisposition :: Maybe String

The content disposition of the entry.

mContentLength :: Integer

The content length of the entry.

mContentMD5 :: Maybe String

The content MD5 of the entry.

mContentType :: Maybe String

The content type of the entry.

mETag :: Maybe String

The ETag of the entry.

mLastModified :: Maybe UTCTime

The last modified time of the entry.

type OpMonad = ReaderT Operator (ExceptT OpenDALError IO)

The monad used for OpenDAL operations.

class Monad m => MonadOperation m where

A type class for monads that can perform OpenDAL operations.

Methods

readOp :: String -> m ByteString

Read the whole path into a bytes.

writeOp :: String -> ByteString -> m ()

Write bytes into given path.

isExistOp :: String -> m Bool

Check if this path exists or not.

createDirOp :: String -> m ()

Create a dir at given path.

copyOp :: String -> String -> m ()

Copy a file from srcPath to dstPath.

renameOp :: String -> String -> m ()

Rename a file from srcPath to dstPath.

deleteOp :: String -> m ()

Delete given path.

statOp :: String -> m Metadata

Get given path’s metadata without cache directly.

listOp :: String -> m Lister

List current dir path. This function will create a new handle to list entries. An error will be returned if path doesn’t end with /.

scanOp :: String -> m Lister

List dir in flat way. Also, this function can be used to list a prefix. An error will be returned if given path doesn’t end with /.

runOp :: Operator -> OpMonad a -> IO (Either OpenDALError a)

Runs an OpenDAL operation in the OpMonad.

newOp :: String -> HashMap String String -> IO (Either OpenDALError Operator)

Creates a new OpenDAL operator via HashMap.

readOpRaw :: Operator -> String -> IO (Either OpenDALError ByteString)

Read the whole path into a bytes.

writeOpRaw :: Operator -> String -> ByteString -> IO (Either OpenDALError ())

Write bytes into given path.

isExistOpRaw :: Operator -> String -> IO (Either OpenDALError Bool)

Check if this path exists or not.

createDirOpRaw :: Operator -> String -> IO (Either OpenDALError ())

Create a dir at given path.

copyOpRaw :: Operator -> String -> String -> IO (Either OpenDALError ())

Copy a file from srcPath to dstPath.

renameOpRaw :: Operator -> String -> String -> IO (Either OpenDALError ())

Rename a file from srcPath to dstPath.

deleteOpRaw :: Operator -> String -> IO (Either OpenDALError ())

Delete given path.

statOpRaw :: Operator -> String -> IO (Either OpenDALError Metadata)

Get given path’s metadata without cache directly.

listOpRaw :: Operator -> String -> IO (Either OpenDALError Lister)

List current dir path. This function will create a new handle to list entries. An error will be returned if path doesn’t end with /.

scanOpRaw :: Operator -> String -> IO (Either OpenDALError Lister)

List dir in flat way. Also, this function can be used to list a prefix. An error will be returned if given path doesn’t end with /.

nextLister :: Lister -> IO (Either OpenDALError (Maybe String))

Get next entry path from Lister.